+++ package.json
@@ -9,8 +9,10 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "ammo.js": "github:kripken/ammo.js",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-qr-code": "^2.0.11",
     "react-scripts": "5.0.1",
     "three": "^0.146.0",
     "three-obj-loader": "^1.1.3",
