+++ src/inde.js
@@ -1,0 +1,2 @@
+const arr =[1,2,3,4,5,6,7,8,9];
+
